added the api's to get user and document data and functionality to join, leave and update presence of document room using socket.io in the sequential loader for better user experience

+++ src/services/documentService.js
@@ -16,4 +16,8 @@ const deleteDoc = async (id) => {
   return api.delete("/document/" + id);
 };
 
-export { createDocument, getDocuments, renameDoc, deleteDoc };
+const accessDocument = async (token) => {
+  return api.get("/document/join/" + token);
+};
+
+export { createDocument, getDocuments, renameDoc, deleteDoc, accessDocument };
